Split adapter and entry points into typed per-mode classes
Second chunk of the v2 extraction/loading refactor. The monolithic
WorkerAdapter and single processTask are replaced by typed per-mode
surfaces so a loading worker can no longer call extraction-only methods
(and vice versa), enforced at compile time.

State:
- BaseState<C,S> (base-state.ts) owns the shared lifecycle: connector vs.
  SDK state, fetch/init/post, the v1->v2 migration shim, and the
  snap-in-version-gated initial domain mapping install (kept shared so a
  loading run still installs the mapping if extraction has not).
- ExtractionState/LoadingState narrow sdkState to ExtractionSdkState /
  LoadingSdkState; createExtractionState/createLoadingState factories.
  Extraction-window resolution is extraction-only. SdkState splits into
  BaseSdkState + ExtractionSdkState + LoadingSdkState.

Adapter:
- BaseAdapter<C,S> owns emit() as a template method with beforeEmit /
  buildEmitPayload / afterEmit hooks. ExtractionAdapter owns repos,
  artifacts, attachment streaming, and boundary updates + repo upload in
  beforeEmit. LoadingAdapter owns load*/mappers/reports. emit payload is
  still gated by event type to preserve exact behavior.
- worker-adapter.helpers.ts renamed to loading-adapter.helpers.ts.

Entry points:
- processExtractionTask / processLoadingTask each build their own typed
  adapter via a shared runWorkerTask driver. Generic
  WorkerAdapterInterface / TaskAdapterInterface / ProcessTaskInterface.
  Attachment processor function types tighten to ExtractionAdapter.
- index.ts: drop processTask/WorkerAdapter, add the four new symbols.

Backwards-compatibility suite intentionally fails on the removed exports
and changed types; baseline regenerated in the final chunk. Default +
timeout-handling suites green.
