+++ src/index.ts
@@ -9,9 +9,13 @@ export type {
   RetryConfig,
   RouteConfig,
 } from './mock-server/mock-server.interfaces';
-export { processTask } from './multithreading/process-task';
+export {
+  processExtractionTask,
+  processLoadingTask,
+} from './multithreading/process-task';
 export { spawn } from './multithreading/spawn/spawn';
-export { WorkerAdapter } from './multithreading/worker-adapter/worker-adapter';
+export { ExtractionAdapter } from './multithreading/worker-adapter/extraction-adapter';
+export { LoadingAdapter } from './multithreading/worker-adapter/loading-adapter';
 export { createMockEvent, MOCK_SERVER_DEFAULT_URL } from './common/test-utils';
 export type { DeepPartial } from './common/test-utils';
 export * from './types';
