+++ src/multithreading/worker-adapter/loading-adapter.helpers.ts
