+++ src/state/state.post-state.test.ts
@@ -1,7 +1,7 @@
 import { mockServer } from '../tests/jest.setup';
 import { createMockEvent } from '../common/test-utils';
 import { EventType } from '../types/extraction';
-import { State, createAdapterState } from './state';
+import { ExtractionState, createExtractionState } from './extraction-state';
 
 /* eslint-disable @typescript-eslint/no-require-imports */
 
@@ -14,8 +14,8 @@ describe('State.postState', () => {
     jest.clearAllMocks();
     jest.restoreAllMocks();
 
-    postStateSpy = jest.spyOn(State.prototype, 'postState');
-    fetchStateSpy = jest.spyOn(State.prototype, 'fetchState');
+    postStateSpy = jest.spyOn(ExtractionState.prototype, 'postState');
+    fetchStateSpy = jest.spyOn(ExtractionState.prototype, 'fetchState');
     processExitSpy = jest.spyOn(process, 'exit').mockImplementation(() => {
       throw new Error('process.exit called');
     });
@@ -37,7 +37,7 @@ describe('State.postState', () => {
 
     postStateSpy.mockRestore();
 
-    const adapterState = await createAdapterState({
+    const adapterState = await createExtractionState({
       event,
       initialState: {},
       initialDomainMapping: {},
@@ -74,7 +74,7 @@ describe('State.postState', () => {
 
     postStateSpy.mockRestore();
 
-    const adapterState = await createAdapterState({
+    const adapterState = await createExtractionState({
       event,
       initialState: {},
       initialDomainMapping: {},
