+++ src/state/state.time-value-resolution.test.ts
@@ -1,7 +1,7 @@
 import { mockServer } from '../tests/jest.setup';
 import { createMockEvent } from '../common/test-utils';
 import { EventType, TimeValue, TimeValueType } from '../types/extraction';
-import { State, createAdapterState } from './state';
+import { ExtractionState, createExtractionState } from './extraction-state';
 
 describe('State — TimeValue resolution', () => {
   let fetchStateSpy: jest.SpyInstance;
@@ -11,7 +11,7 @@ describe('State — TimeValue resolution', () => {
     jest.clearAllMocks();
     jest.restoreAllMocks();
 
-    fetchStateSpy = jest.spyOn(State.prototype, 'fetchState');
+    fetchStateSpy = jest.spyOn(ExtractionState.prototype, 'fetchState');
     processExitSpy = jest.spyOn(process, 'exit').mockImplementation(() => {
       throw new Error('process.exit called');
     });
@@ -40,7 +40,7 @@ describe('State — TimeValue resolution', () => {
 
       // Act & Assert
       await expect(
-        createAdapterState({
+        createExtractionState({
           event,
           initialState: {},
           initialDomainMapping: {},
@@ -74,7 +74,7 @@ describe('State — TimeValue resolution', () => {
 
       // Act & Assert
       await expect(
-        createAdapterState({
+        createExtractionState({
           event,
           initialState: {},
           initialDomainMapping: {},
@@ -109,7 +109,7 @@ describe('State — TimeValue resolution', () => {
       fetchStateSpy.mockResolvedValue({ state: stringifiedState });
 
       // Act
-      const state = await createAdapterState({
+      const state = await createExtractionState({
         event,
         initialState: {},
         initialDomainMapping: {},
@@ -150,7 +150,7 @@ describe('State — TimeValue resolution', () => {
       fetchStateSpy.mockResolvedValue({ state: stringifiedState });
 
       // Act
-      await createAdapterState({
+      await createExtractionState({
         event,
         initialState: {},
         initialDomainMapping: {},
@@ -189,7 +189,7 @@ describe('State — TimeValue resolution', () => {
       fetchStateSpy.mockResolvedValue({ state: stringifiedState });
 
       // Act
-      await createAdapterState({
+      await createExtractionState({
         event,
         initialState: {},
         initialDomainMapping: {},
